Rename parser -> Parser and formats -> FORMATS in apachelog.parser.
This brings us closer to PEP8_ compliance (CapWords class names).
Global constants are not given an explict case convention in PEP8, but
UPPERCASE seems conventional (for example, see the `os` module).

I also moved the main docstring and metadata from
`apachelog.parser.__doc__` to `apachelog.__doc__`, so it will be
easier to find.  I updated that docstring to use reStructuredText, and
converted the examples to doctests.

.. _PEP8: http://www.python.org/dev/peps/pep-0008/